Daedalus.Cost

Documentation

class Modality m => Costly m where

A generic concept of cost functions and heuristic for modalities, adapted from the function of the same name in monad-dijkstra package.

Associated Types

type CostOf m :: *

Methods

cost :: CostOf m -> CostOf m -> StrategySearchT m ()