Daedalus.SearchT

Documentation

type Modality = Monad

newtype SearchT m t

Constructors

SearchT

Fields

unSearchT :: forall tech. Monoid tech => ReaderT (ContT tech m tech) (ContT tech (ContT tech m)) t

techCallcc2 :: (MonadTrans t, MonadReader r1 (t (ContT r m)), MonadCont m) => ((r -> m b1) -> ReaderT r1 (ContT r m) b) -> t (ContT r m) b

scope :: Modality m => SearchT m t -> SearchT m t

scope and cut work similarly as the seal and collapse functions respectively from monad-dijkstra. cut works similarly as Prolog's cut operation but uses scope blocks to control how far the cut extends.

cut :: Modality m => SearchT m a

techResolveCallback :: forall m t. Modality m => (forall tech. SearchT m tech -> SearchT m tech) -> SearchT m t -> SearchT m t

techResolveCallback is a tool to make sure that polymorphic transformations operate over enough of the program, so that left-distributivity can be satisfied.

techExpHoistContT :: (m r -> n r2) -> (n r2 -> m r) -> ContT r m t -> ContT r2 n t

techExpHoistSearchT :: (forall t. m t -> n t) -> (forall t. n t -> m t) -> SearchT m t -> SearchT n t

SearchT enjoys exponential maps in the category of functors.

techCoda :: (forall tech. Monoid tech => tech -> ContT tech m tech) -> SearchT m t -> SearchT m t

techCoda applies a function to run following the operation chronologically. It can be used to do cleanup on a branch.

runSearchT :: Modality m => SearchT m () -> m ()

A basic eliminator for SearchT modality transformer.

collectSearchT :: Modality m => SearchT (WriterT [t] m) t -> m [t]

Example:

>>> collectSearchT(scope(return(1::Int)<|>cut<|>return(2::Int))<|> return 3)
[1,3]