Daedalus.Strategy.AStar

Documentation

data AStarT c m t

The search tree has to be partially reified to achieve the effect of a AStar search. You generally do find something algebraic; sometimes it's a bit strange.

aStarStrategy :: (Modality m, Monoid c, Ord c) => SearchT (AStarT c m) t -> SearchT (AStarT c m) t

The strategy to use on disjunctions just captures the parameter and puts it in a priority queue for later.

runAStarT :: forall m c. (Modality m, Monoid c, Ord c, Bounded c) => StrategySearchT (AStarT c m) () -> m ()

runAStarT operator does much of the work for the A* strategy.

Example:

>>> runAStarT$((cost(5::Sum Int) 0>>applyStrategy(return 'A'))<||>((cost 11 0>>applyStrategy(return 'B'))<||>(cost 3 0>>applyStrategy(return 'C'))))>>=lift.lift.print
'C'
'A'
'B'

Nota bene that with this strategy it is important to apply applyStrategy in the leaves of the user code computation.