Daedalus.StrategySearch

Documentation

data StrategySearchT m t

StrategySearchT allows to insert a (polymorphic) strategy combinator at every choice point in a search.

strategySearchT :: Modality m => SearchT m t -> StrategySearchT m t

It lifts a regular old SearchT, necessarily without applying any strategies within it.

applyStrategy :: Modality m => StrategySearchT m t -> StrategySearchT m t

(<||>) :: Monad m => StrategySearchT m a -> StrategySearchT m a -> StrategySearchT m a infixl 3

Convenience combinator for applying a strategy to both arms of a disjunction.

expHoistStrategySearch :: (Modality m, Modality n) => (forall t. m t -> n t) -> (forall t. n t -> m t) -> StrategySearchT m t -> StrategySearchT n t

addStrategy :: Modality m => (forall u. SearchT m u -> SearchT m u) -> StrategySearchT m t -> StrategySearchT m t

Composes a strategy into the strategy chain.

runStrategySearchT :: Modality m => StrategySearchT m t -> SearchT m t

A basic eliminator into the SearchT modality transformer.

module Daedalus.SearchT